requirements.txt
## The following requirements were added by pip freeze:
alembic==0.8.8
amqp==2.1.1
anyjson==0.3.3
astroid==1.4.8
Babel==2.3.4
bcrypt==3.1.1
beautifulsoup4==4.5.1
billiard==3.5.0.2
blinker==1.3
git+https://github.com/Atheloses/cachecore.git
celery==3.1.0
celery-with-redis==3.0
cffi==1.8.3
click==6.6
coverage==3.6
coveralls==0.3
cssmin==0.2.0
docopt==0.6.1
fake-factory==0.7.2
flake8==2.0
Flask==0.11.1
Flask-Assets==0.12
Flask-Babel==0.11.1
Flask-Bcrypt==0.7.1
Flask-Heroku-Env==0.0.4
Flask-Login==0.4.0
Flask-Mail==0.9.0
Flask-Migrate==2.0.0
Flask-Restless==0.17.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-Travis==0.0.2
Flask-WTF==0.9.2
gunicorn==18.0
isort==4.2.5
itsdangerous==0.23
Jinja2==2.8
jsmin==2.0.5
kombu==3.0.30
lazy-object-proxy==1.2.2
Mako==1.0.5
MarkupSafe==0.18
mccabe==0.2.1
mimerender==0.6.0
pep8==1.4.6
psutil==4.4.2
psycopg2==2.5.1
py==1.4.31
py-bcrypt==0.4
pycparser==2.17
pycrypto==2.6.1
pyflakes==0.7.3
pylint==1.6.4
pyScss==1.2.0
pytest==2.5.2
python-dateutil==2.1
python-editor==1.0.1
python-mimeparse==1.6.0
pytz==2016.7
PyYAML==3.10
redis==2.10.5
requests==2.4.3
sh==1.9
six==1.4.1
SQLAlchemy==1.1.3
vine==1.1.3
waitress==1.0.1
webassets==0.12.0
WebOb==1.6.2
WebTest==2.0.15
Werkzeug==0.11.11
wheel==0.24.0
wrapt==1.10.8
WTForms==2.1
git+https://github.com/mikeboers/Flask-ACL.git
git+https://github.com/Atheloses/flask-social.git
pyvat
rauth==0.7.0
